requirements.txt
langchain
streamlit
openai
python-decouple
langchain-openai
langchain-community